refactor(constants): Move TOOL_CONNECTION_MODELS to a new constants file and clean up imports in Slack services
