+++ src/slack/constants.ts
@@ -1,0 +1,21 @@
+import {
+  GithubConfig,
+  HubspotConfig,
+  JiraConfig,
+  LinearConfig,
+  McpConnection,
+  NotionConfig,
+  PostgresConfig,
+  SalesforceConfig
+} from '../database/models';
+
+export const TOOL_CONNECTION_MODELS = [
+  GithubConfig,
+  HubspotConfig,
+  JiraConfig,
+  LinearConfig,
+  McpConnection,
+  NotionConfig,
+  PostgresConfig,
+  SalesforceConfig
+];
